+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Support for cross compilation for `musl`
 
+### Changed
+* Rewrite of nstar for a better user experience
+
 ## [0.5.0] - 2020-11-2
 ### Added
 * Notifications for system events are now sent to connected clients (e.g. to `nstar`)